3GPP stop-gap version
The problem-details specification has been parked for a while waiting
for CoRAL to complete.
In the meantime, 3GPP has written the problem-details specification
into their client-server specifications and is expecting to cut a
release in June 2022.
It is unlikely that the CoRAL is complete by then, which would cause
problem-details to be written out of the specification (and probably
be replaced by an ad-hoc mechanism).

This PR proposes to define a concise problem details structure that
is simple enough so it can be completed within a couple of months.
This would keep the problem details approach within the 3GPP system
and will allow us to add a more complete CoRAL version later.